[Refactor] #5 User.java 경로 이동
